requirements.txt
farm-haystack -f https://download.pytorch.org/whl/torch_stable.html
farm-haystack[faiss]
sentence-transformers